Create a passive environment, for comparison.
Create a passive hawk driver and trans chaining sequences.

The passive driver pulls items but does not touch the interface.
It does this so that the existing chaining sequences can continue
to work.

The trans chaining sequence is necessary because otherwise it will
try to fetch tags and unite them with responses, causing errors.